LayoutTests/fast/viewport/viewport-46.html

<head>
    <title>Viewport META with "minimum-scale=11, maximum-scale=13, initial-scale=12".</title>
    <meta name="viewport" content="minimum-scale=11, maximum-scale=13, initial-scale=12">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
