LayoutTests/fast/viewport/viewport-46-expected.txt

CONSOLE MESSAGE: line 3: Viewport maximum-scale cannot be larger than 10.0. The maximum-scale will be set to 10.0.
CONSOLE MESSAGE: line 3: Viewport maximum-scale cannot be larger than 10.0. The maximum-scale will be set to 10.0.
CONSOLE MESSAGE: line 3: Viewport maximum-scale cannot be larger than 10.0. The maximum-scale will be set to 10.0.
viewport size 32x35 scale 10.000000 with limits [10.000000, 10.000000]
